Crank all members of struct ipipstat to 64 bits
Will make conversion to percpu counters easier.  ok bluhm@

+++ sys/netinet/ip_ipip.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: ip_ipip.h,v 1.6 2007/12/14 18:33:41 deraadt Exp $ */
+/*	$OpenBSD: ip_ipip.h,v 1.7 2017/02/20 17:04:25 jca Exp $ */
 /*
  * The authors of this code are John Ioannidis (ji@tla.org),
  * Angelos D. Keromytis (kermit@csd.uch.gr) and
@@ -43,18 +43,17 @@
  * Not quite all the functionality of RFC-1853, but the main idea is there.
  */
 
-struct ipipstat
-{
-    u_int32_t	ipips_ipackets;		/* total input packets */
-    u_int32_t	ipips_opackets;		/* total output packets */
-    u_int32_t	ipips_hdrops;		/* packet shorter than header shows */
-    u_int32_t	ipips_qfull;
-    u_int64_t   ipips_ibytes;
-    u_int64_t   ipips_obytes;
-    u_int32_t	ipips_pdrops;		/* packet dropped due to policy */
-    u_int32_t	ipips_spoof;		/* IP spoofing attempts */
-    u_int32_t   ipips_family;		/* Protocol family mismatch */
-    u_int32_t   ipips_unspec;            /* Missing tunnel endpoint address */
+struct ipipstat {
+    u_int64_t	ipips_ipackets;		/* total input packets */
+    u_int64_t	ipips_opackets;		/* total output packets */
+    u_int64_t	ipips_hdrops;		/* packet shorter than header shows */
+    u_int64_t	ipips_qfull;
+    u_int64_t	ipips_ibytes;
+    u_int64_t	ipips_obytes;
+    u_int64_t	ipips_pdrops;		/* packet dropped due to policy */
+    u_int64_t	ipips_spoof;		/* IP spoofing attempts */
+    u_int64_t	ipips_family;		/* Protocol family mismatch */
+    u_int64_t	ipips_unspec;            /* Missing tunnel endpoint address */
 };
 
 #define IP4_DEFAULT_TTL    0
